pyproject.toml
[project]
name = "30122-project-asia-pacific-enthusiastic-coding-apec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=2.18.2",
    "dash-bootstrap-components>=1.7.1",
    "httpx>=0.28.1",
    "ipython>=8.32.0",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "plotly-express>=0.4.1",
    "pycountry>=24.6.1",
]
